requirements.txt
numpy==1.26.4
setuptools==69.5.1
git+https://github.com/ai-forever/Kandinsky-2.git@33c3176
git+https://github.com/openai/CLIP.git@a1d0717
diffusers==0.30.3
git+https://github.com/huggingface/transformers@6544271  
torchao==0.7.0
optimum-quanto==0.2.6
mmgp==1.2.0
tokenizers==0.20.1
accelerate==1.0.1
safetensors==0.4.5
opencv-python==4.8.0.74
gradio==3.50.2
toolz==1.0.0
psutil==5.9.5
deepdiff==8.0.1
torchsde==0.2.6
compel==2.0.3
huggingface-hub==0.25.2
https://github.com/jllllll/bitsandbytes-windows-webui/releases/download/wheels/bitsandbytes-0.41.2.post2-py3-none-win_amd64.whl; platform_system == "Windows"
bitsandbytes==0.41.1; platform_system != "Windows"